Made layout improvements.
Seperate the calculation of horisontal and vertical offset of nodes. This was nesesary in order to calculate hights based on content ...a variable not set at the creation of a given node. Many of these details effect block and inline-block nodes. There was great complication with the positioning of incomplete rows. That has also been fixed. All of this new code could stand to be cleaned up!