Extensive flow-controll and other changes
I changed the code in too many way to sumarize. Unfortunately, there is still one issue to be resolved: somehow the elements are being calculated and drawn multiple times. This slows everything down and also produces display artifacts. It would seem that part or all of the Render Tree is being duplicated.